refactor: 검색 필터 필드 분리 #2
